backend/requirements.txt
flask
flask_cors
flask-socketio
redis
psycopg2-binary
eventlet